refactor(api): UserSearchService to UserCreateService
