+++ alarmeet-api/src/main/java/click/alarmeet/alarmeetapi/apis/groups/usecase/GroupUseCase.java
@@ -12,7 +12,7 @@ import click.alarmeet.alarmeetapi.apis.groups.mapper.GroupMapper;
 import click.alarmeet.alarmeetapi.apis.groups.mapper.GroupUserMapper;
 import click.alarmeet.alarmeetapi.apis.groups.service.GroupSaveService;
 import click.alarmeet.alarmeetapi.apis.groups.service.GroupSearchService;
-import click.alarmeet.alarmeetapi.apis.users.service.UserSearchService;
+import click.alarmeet.alarmeetapi.apis.users.service.UserCreateService;
 import click.alarmeet.alarmeetapi.common.annotation.UseCase;
 import click.alarmeet.alarmeetcommon.exception.GlobalErrorException;
 import lombok.RequiredArgsConstructor;
@@ -28,7 +28,7 @@ public class GroupUseCase {
 	private final GroupSaveService groupSaveService;
 	private final GroupSearchService groupSearchService;
 
-	private final UserSearchService userSearchService;
+	private final UserCreateService userCreateService;
 
 	public void createGroup(String userId, GroupCreateReq groupReq) {
 		ObjectId userOid = new ObjectId(userId);
@@ -45,6 +45,6 @@ public class GroupUseCase {
 			)
 		);
 
-		userSearchService.addGroupId(userOid, group.getId());
+		userCreateService.addGroupId(userOid, group.getId());
 	}
 }
